[rv_plic] Move exemplary parametrization to a separate directory
rv_plic is heavily customizable by generating some parts of the design
form templates. Currently the hw/ip/rv_plic directory contains "common"
files which are always used and not templated, and some files which are
produced from templates.

More clearly those two types of files by moving all generated files into
an example directory, indicating that these files are an exemplary
parametrization of the IP core which can be used for IP-level DV and lint
purposes, but isn't used as-is in e.g. top_earlgrey.

There are no functional changes from this move, the files are kept the
same.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>